refactor(ses): extract SesSuppressionService from SesService (strangler step 6)
Sixth store-based domain split, resolving the shared-helper
deferral from the account step: account suppression and the suppression list
share validateSuppressionReason, so both stores move together into one
@ApplicationScoped SesSuppressionService where the helper becomes private.
The send filters stay in the facade — getEffectiveSuppressedReasons reads a
configuration set's options or falls back here, and
collectSuppressedReasons/resolveSuppressionReason read entries back through
findSuppressedDestination. Consumers (v1 SesQueryHandler, v2 SesController)
are unchanged.
