+++ src/test/java/io/github/hectorvent/floci/services/ses/SesServiceSuppressionLegacyKeyTest.java
@@ -51,8 +51,7 @@ class SesServiceSuppressionLegacyKeyTest {
                 new SesAccountService(new InMemoryStorage<String, Boolean>()),
                 new InMemoryStorage<String, EmailTemplate>(),
                 new InMemoryStorage<String, ConfigurationSet>(),
-                suppressionStore,
-                new InMemoryStorage<String, AccountSuppressionAttributes>(),
+                new SesSuppressionService(suppressionStore, new InMemoryStorage<String, AccountSuppressionAttributes>()),
                 new InMemoryStorage<String, AccountVdmAttributes>(),
                 new InMemoryStorage<String, DedicatedIpPool>(),
                 new SesContactService(new InMemoryStorage<String, ContactList>(), new InMemoryStorage<String, Contact>(), Clock.systemUTC()),
