+++ src/test/java/io/github/hectorvent/floci/services/ses/SesServiceSuppressionReasonTest.java
@@ -51,8 +51,7 @@ class SesServiceSuppressionReasonTest {
                 new SesAccountService(new InMemoryStorage<String, Boolean>()),
                 new InMemoryStorage<String, EmailTemplate>(),
                 new InMemoryStorage<String, ConfigurationSet>(),
-                suppressionStore,
-                accountSuppressionStore,
+                new SesSuppressionService(suppressionStore, accountSuppressionStore),
                 new InMemoryStorage<String, AccountVdmAttributes>(),
                 new InMemoryStorage<String, DedicatedIpPool>(),
                 new SesContactService(new InMemoryStorage<String, ContactList>(), new InMemoryStorage<String, Contact>(), Clock.systemUTC()),
